third_party/blink/web_tests/fast/forms/week/ValidityState-typeMismatch-week-expected.txt

This test aims to check for typeMismatch flag with type=week input fields

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS "" is a correct valid week string.
PASS "0001-W01" is a correct valid week string.
PASS "1583-W01" is a correct valid week string.
PASS "9999-W52" is a correct valid week string.
PASS "275760-W37" is a correct valid week string.
PASS "2009-W01" is a correct valid week string.
PASS "2004-W53" is a correct valid week string.
PASS "2003-W52" is a correct valid week string.
PASS "1992-W53" is a correct valid week string.
PASS " 2009-W01 " is an invalid week string and was sanitized.
PASS "2009W01" is an invalid week string and was sanitized.
PASS "2009-w01" is an invalid week string and was sanitized.
PASS "2009-W1" is an invalid week string and was sanitized.
PASS "2009-W001" is an invalid week string and was sanitized.
PASS "a" is an invalid week string and was sanitized.
PASS "-1-W09" is an invalid week string and was sanitized.
PASS "0000-W52" is an invalid week string and was sanitized.
PASS "2147483648-W01" is an invalid week string and was sanitized.
PASS "275760-W38" is an invalid week string and was sanitized.
PASS "2009-W00" is an invalid week string and was sanitized.
PASS "2009-W-1" is an invalid week string and was sanitized.
PASS "2004-W54" is an invalid week string and was sanitized.
PASS "2003-W53" is an invalid week string and was sanitized.
PASS "1992-W54" is an invalid week string and was sanitized.
PASS "2009/09" is an invalid week string and was sanitized.
PASS "200909" is an invalid week string and was sanitized.
PASS "2009-Wxx" is an invalid week string and was sanitized.
PASS "2009" is an invalid week string and was sanitized.
PASS "invalid" is an invalid week string and was sanitized while disabled.
PASS successfullyParsed is true

TEST COMPLETE
